+++ CHANGELOG.md
@@ -3,14 +3,16 @@
 ### Added
 
 - Add reset button in pdf output view
+- Add data, regione, idroecoregione, area, bacino to log
 
 ### Changed
 
 - Drop panel for output in PDF view
 - Drop "change view" button
 - Split `core::csv::{deser, parser}` into niseci and hfbi
-- Round x1, x2, x3 to 3 decimal digits
-- Round RQE Niseci to 2 decimal digits
+- Round `x1`, `x2`, `x3` to 3 decimal digits
+- Round `RQE Niseci` to 2 decimal digits
+- Update `ComunitaNISECI` format
 
 ## [0.0.26] - 2025-09-16